# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "DataPorts API with swagger",
    "description": "This is a simple REST API made with Node.JS and documented with Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "José A. Clemente",
      "email": "jclemente@prodevelop.es"
    }
  },
  "basePath": "/dac",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/images/{id}/template": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "getTemplate",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Returns a template for creating an agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to retrieve its template",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
    "/agents/ngsiagent": {
      "get": {
        "tags": [
          "NGSI Agent"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Returns all agents on the system",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      },
      "post": {
        "tags": [
          "NGSI Agent"
        ],
        "summary": "Creates and starts an agent given a template",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ngsiagentTemplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/agents/ngsiagent/{id}": {
      "delete": {
        "tags": [
          "NGSI Agent"
        ],
        "summary": "Deletes an NGSI Agent",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the agent that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/agents/ngsiagent/{id}/start": {
      "patch": {
        "tags": [
          "NGSI Agent"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the agent that we want to start",
            "type": "string"
          }
        ],
        "summary": "Starts the specified agent on the url",
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/agents/ngsiagent/{id}/stop": {
      "patch": {
        "tags": [
          "NGSI Agent"
        ],
        "operationId": "stop",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the agent that we want to stop",
            "type": "string"
          }
        ],
        "summary": "Stops the specified agent on the url",
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/agents/ngsiagent/{id}/log": {
      "get": {
        "tags": [
          "NGSI Agent"
        ],
        "operationId": "log",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the agent that we want to recover its log",
            "type": "string"
          }
        ],
        "summary": "Return the logs of the specified agent on the url",
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/agents/ngsiagent/{id}/inspect": {
      "get": {
        "tags": [
          "NGSI Agent"
        ],
        "operationId": "inspect",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the agent that we want to inspect",
            "type": "string"
          }
        ],
        "summary": "Returns de most relevant information of an agent",
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
    "/dataModel": {
      "get": {
        "tags": [
          "DataModel"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Retrieve known datamodels",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "succesful operation"
          }
        }
      }
    },
    "/dataModel/getSchema": {
      "post": {
        "tags": [
          "DataModel"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Register a new datamodel",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Creates a datamodel in the database",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token with the `Bearer` prefix, E.g., \"Bearer easkkassjejm.dadsaeea.ssdaee\""
    }
  },
  "definitions": {
...
    "GenericMessage": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "GenericAnswer": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ngsiagentTemplate": {
      "properties": {
        "ContainerName": {
          "type": "string"
        },
        "Hostname": {
          "type": "string"
        },
        "Domainname": {
          "type": "string"
        },
        "User": {
          "type": "string"
        },
        "Image": {
          "type": "string"
        },
        "Env": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}